# utils: managed_bytes: optimize memory usage for small buffers