--- a/utils/managed_bytes.cc
+++ b/utils/managed_bytes.cc
@@ -27,18 +27,6 @@ managed_bytes_opt to_managed_bytes_opt(const bytes_opt& bo) {
     return managed_bytes(*bo);
 }
 
-std::unique_ptr<bytes_view::value_type[]>
-managed_bytes::do_linearize_pure() const {
-    auto b = _u.ptr;
-    auto data = std::unique_ptr<bytes_view::value_type[]>(new bytes_view::value_type[b->size]);
-    auto e = data.get();
-    while (b) {
-        e = std::copy_n(b->data, b->frag_size, e);
-        b = b->next;
-    }
-    return data;
-}
-
 sstring to_hex(const managed_bytes& b) {
     return fmt::to_string(managed_bytes_view(b));
 }