@dorlaor 2023-06-15
Wow

On Thu, Jun 15, 2023 at 5:41 PM Michał Chojnowski ***@***.***>
wrote:

> utils: managed_bytes: optimize memory usage for small buffers
>
> managed_bytes is implemented as chain of blob_storage objects.
> Each blob_storage contains 24 bytes of metadata. But in the most
> common case -- when there is only a single element in the chain --
> 16 bytes of this metadata is trivial/unused.
>
> This is regrettable waste because managed_bytes is used for every
> database cell in the memtables and cache. It means that every value
> of size >= 7 bytes (smaller ones fit in the inline storage of
> managed_bytes) receives 16 bytes of useless overhead.
>
> To correct that, this series adds to managed_bytes an alternative storage
> layout -- used for buffers small enough to fit in one fragment -- which
> only
> stores the necessary minimum of metadata. (That is: a pointer to the
> parent,
> to facilitate moving the storage during memory defragmentation).
>
> This saves 16 bytes on every cell greater than 15 bytes. Which includes
> e.g.
> every live cell with value bigger than 6 bytes, which likely applies to
> most cells.
> ------------------------------
> You can view, comment on, or merge this pull request online at:
>
>   https://github.com/scylladb/scylladb/pull/14263
> Commit Summary
>
>    - 70a9514
>    <https://github.com/scylladb/scylladb/pull/14263/commits/70a9514384267936324f2527819e928fac7fdbc6>
>    utils: managed_bytes: rewrite managed_bytes methods in terms of
>    managed_bytes_view
>    - d2cf4d5
>    <https://github.com/scylladb/scylladb/pull/14263/commits/d2cf4d5c40958c5bf15f6109df59e01900dae991>
>    utils: managed_bytes: optimize memory usage for small buffers
>    - ee96e47
>    <https://github.com/scylladb/scylladb/pull/14263/commits/ee96e47bd73d614a771972adf71dbf170be333c3>
>    utils: managed_bytes: optimize memory usage for small buffers
>
> File Changes
>
> (3 files <https://github.com/scylladb/scylladb/pull/14263/files>)
>
>    - *M* utils/managed_bytes.cc
>    <https://github.com/scylladb/scylladb/pull/14263/files#diff-7f2b4533cad6444c6934cc58dcb997024943b9b18e7f420604ad539d8771d001>
>    (12)
>    - *M* utils/managed_bytes.hh
>    <https://github.com/scylladb/scylladb/pull/14263/files#diff-5d4d64bcc5a9e030a442609b7639e8c33475e722aa2849a01360b7a91081008d>
>    (387)
>    - *M* utils/mutable_view.hh
>    <https://github.com/scylladb/scylladb/pull/14263/files#diff-313a98675cd5173f099487878d494297ea90e16000ebf994362b05937ea3945d>
>    (3)
>
> Patch Links:
>
>    - https://github.com/scylladb/scylladb/pull/14263.patch
>    - https://github.com/scylladb/scylladb/pull/14263.diff
>
> —
> Reply to this email directly, view it on GitHub
> <https://github.com/scylladb/scylladb/pull/14263>, or unsubscribe
> <https://github.com/notifications/unsubscribe-auth/AANHUROMLKOCFOVVA7A4NFLXLMNKFANCNFSM6AAAAAAZH6Y3JU>
> .
> You are receiving this because you are subscribed to this thread.Message
> ID: ***@***.***>
>