@avikivity 2023-06-16
> It seems that reimplementing the specialized `managed_bytes(const managed_bytes& o)` with the general view-based constructor causes a noticeable regression in `perf_simple_query`. (~43800 instr/op -> ~45200 instr/op). I'll have to take a closer look at why that's the case.

Wow.