@michoecho 2023-06-16
> Looks good.
> 
> We added some interoperability between bytes_ostream and managed_bytes, does it need adjustment?

No, bytes_ostream just unconditionally uses the multi-chunk layout (which is the same as before, so bytes_ostream logic doesn't have to change), even if it there is only a single chunk.
IIUC the output of bytes_ostream is never stored long-term in LSA, so the waste of several bytes there isn't worth improving.