@michoecho 2023-06-21
v2:
- fixed typos in comments
- hoisted `inline_size` outside the union
- removed padding fields in the union variants
- ditched wrapper structs over union variants
- brought back fast paths (for the small variants) for the copy constructor and operator[] of managed_bytes, because they matter in microbenchmarks
- renamed "small, medium, big" to "inline, single-chunk, multi-chunk" everywhere

> > It seems that reimplementing the specialized `managed_bytes(const managed_bytes& o)` with the general view-based constructor causes a noticeable regression in `perf_simple_query`. (~43800 instr/op -> ~45200 instr/op). I'll have to take a closer look at why that's the case.
> 
> Wow.

I didn't dig deep, but from what I gathered, the difference had mostly to do with copying clustering keys, which in `perf_simple_query` fit into the inline storage, so they benefited from the fast path which only needs like 4 instructions. (I checked that by growing the keys to 16 bytes — it made the difference go away.)
The difference seems excessive (is the general path so expensive, or are the keys copied so much?) but I didn't want to spend too much time on this — it's not that interesting — so I just brought back the fast path.