@michoecho 2023-06-21
> CI state **`FAILURE`** - https://jenkins.scylladb.com/job/scylla-master/job/scylla-ci/1988/

I got defeated by a C++ gotcha here — I forgot that when delegating constructors are used, the object is considered constructed as soon as the delegating constructor finishes. But I used a delegating constructor in the copy constructor I just added, and when the body of the copy constructor threw (in a bad_alloc test), the destructor was called unexpectedly (for me).