@michoecho 2023-06-22
> CI state FAILURE - https://jenkins.scylladb.com/job/scylla-master/job/scylla-ci/1992/

And here I was just sloppy. The newest force-push should be correct.