@avikivity 2023-06-22
Please also add memory_footprint_test report to the cover letter.

No, I don't mind if the cover letter is much larger than the patch.