@mykaul 2023-08-03
> > @michoecho - can you push this forward, so we'll get it into 5.4?
> 
> I could, but do we really want it? The improvement is minor. Maybe we shouldn't do any minor improvements until we are comfortable with stability?

It can certainly wait post 5.4 branching then.