@michoecho 2024-01-11
> ping, can we merge this change?

We could, it's just that the change is lackluster. We can do better.

Every cell bigger than 15 bytes (which includes every live cell where the value is bigger than 6 bytes) has 40 bytes of overhead. This PR lowers the overhead to 24 bytes. But what we really want is ~2 bytes, which you can get with a data structure which groups multiple cells within one allocation. It shouldn't be very hard to code, we just never got to it.

Note: I just noticed that the patch is missing a relevant scylla-gdb.py update.