@mykaul 2024-01-11
> We could, it's just that the change is lackluster. We can do better.

Don't let Perfect be the enemy of good - let's get it in!