@denesb 2024-01-12
@michoecho please rebase, address https://github.com/scylladb/scylladb/pull/14263#issuecomment-1602347945 and this PR is good to go. You can always improve on it in a follow-up.